requirements.txt
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
daphne==4.0.0
Django==4.2.7
django-celery-results==2.5.1
django-redis==5.4.0
djangorestframework==3.14.0
docker==6.1.3
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.4
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv
pytz==2023.3.post1
